@@ -1,11 +1,20 @@
 module github.com/alexboor/lbx-telebot
 
-go 1.20
-
-require gopkg.in/telebot.v3 v3.1.3
+go 1.21
 
 require (
+	github.com/go-resty/resty/v2 v2.10.0
+	github.com/jackc/pgx/v4 v4.18.1
+	github.com/jdkato/prose/v2 v2.0.0
+	github.com/wcharczuk/go-chart/v2 v2.1.1
+	golang.org/x/exp v0.0.0-20230801115018-d63ba01acd4b
+	gopkg.in/telebot.v3 v3.1.3
+)
+
+require (
+	github.com/blend/go-sdk v1.20220411.3 // indirect
 	github.com/deckarep/golang-set v1.7.1 // indirect
+	github.com/golang/freetype v0.0.0-20170609003504-e2365dfdc4a0 // indirect
 	github.com/jackc/chunkreader/v2 v2.0.1 // indirect
 	github.com/jackc/pgconn v1.14.0 // indirect
 	github.com/jackc/pgio v1.0.0 // indirect
@@ -13,13 +22,12 @@ require (
 	github.com/jackc/pgproto3/v2 v2.3.2 // indirect
 	github.com/jackc/pgservicefile v0.0.0-20221227161230-091c0ba34f0a // indirect
 	github.com/jackc/pgtype v1.14.0 // indirect
-	github.com/jackc/pgx/v4 v4.18.1 // indirect
 	github.com/jackc/puddle v1.3.0 // indirect
-	github.com/jdkato/prose/v2 v2.0.0 // indirect
 	github.com/mingrammer/commonregex v1.0.1 // indirect
-	golang.org/x/crypto v0.6.0 // indirect
-	golang.org/x/exp v0.0.0-20230801115018-d63ba01acd4b // indirect
-	golang.org/x/text v0.7.0 // indirect
+	golang.org/x/crypto v0.14.0 // indirect
+	golang.org/x/image v0.11.0 // indirect
+	golang.org/x/net v0.17.0 // indirect
+	golang.org/x/text v0.13.0 // indirect
 	gonum.org/v1/gonum v0.7.0 // indirect
 	gopkg.in/neurosnap/sentences.v1 v1.0.6 // indirect
 )
