@@ -1,7 +1,6 @@
-package handler
+package model
 
 import (
-	"github.com/alexboor/lbx-telebot/internal/model"
 	"reflect"
 	"testing"
 	"time"
@@ -119,37 +118,37 @@ func Test_parseSingleOpt(t *testing.T) {
 	tests := []struct {
 		name string
 		opt  string
-		want model.Option
+		want Option
 		ok   bool
 	}{
 		{
 			name: "string",
 			opt:  "wrong",
-			want: model.Option{},
+			want: Option{},
 			ok:   false,
 		},
 		{
 			name: "incorrect duration 1ss",
 			opt:  "1ss",
-			want: model.Option{},
+			want: Option{},
 			ok:   false,
 		},
 		{
 			name: "incorrect duration 1dd",
 			opt:  "1dd",
-			want: model.Option{},
+			want: Option{},
 			ok:   false,
 		},
 		{
 			name: "correct int",
 			opt:  "15",
-			want: model.Option{Limit: 15},
+			want: Option{Limit: 15},
 			ok:   true,
 		},
 		{
 			name: "correct duration",
 			opt:  "1s",
-			want: model.Option{Date: time.Now().Truncate(time.Hour * 24)},
+			want: Option{Date: time.Now().Truncate(time.Hour * 24)},
 			ok:   true,
 		},
 	}
@@ -170,37 +169,37 @@ func Test_parseTwoOpts(t *testing.T) {
 	tests := []struct {
 		name string
 		opts []string
-		want model.Option
+		want Option
 		ok   bool
 	}{
 		{
 			name: "incorrect limit",
 			opts: []string{"1s", "tess"},
-			want: model.Option{},
+			want: Option{},
 			ok:   false,
 		},
 		{
 			name: "incorrect duration 1st",
 			opts: []string{"1ss", "12"},
-			want: model.Option{},
+			want: Option{},
 			ok:   false,
 		},
 		{
 			name: "incorrect duration 2nd",
 			opts: []string{"12", "1ss"},
-			want: model.Option{},
+			want: Option{},
 			ok:   false,
 		},
 		{
 			name: "correct limit 1st",
 			opts: []string{"12", "1s"},
-			want: model.Option{Limit: 12, Date: time.Now().Truncate(time.Hour * 24)},
+			want: Option{Limit: 12, Date: time.Now().Truncate(time.Hour * 24)},
 			ok:   true,
 		},
 		{
 			name: "correct limit 2nd",
 			opts: []string{"1s", "12"},
-			want: model.Option{Limit: 12, Date: time.Now().Truncate(time.Hour * 24)},
+			want: Option{Limit: 12, Date: time.Now().Truncate(time.Hour * 24)},
 			ok:   true,
 		},
 	}
@@ -217,139 +216,139 @@ func Test_parseTwoOpts(t *testing.T) {
 	}
 }
 
-func Test_parseTopAndBottomPayload(t *testing.T) {
+func TestNewProfileOption(t *testing.T) {
 	tests := []struct {
 		name    string
 		payload string
-		want    model.Option
-		ok      bool
-	}{
-		{
-			name:    "1 duration opt",
-			payload: "1s",
-			want:    model.Option{Date: time.Now().Truncate(time.Hour * 24)},
-			ok:      true,
-		},
-		{
-			name:    "1 limit opt",
-			payload: "3",
-			want:    model.Option{Limit: 3},
-			ok:      true,
-		},
-		{
-			name:    "2 opts",
-			payload: "1s 12",
-			want:    model.Option{Date: time.Now().Truncate(time.Hour * 24), Limit: 12},
-			ok:      true,
-		},
-		{
-			name:    "2 opts duration",
-			payload: "1s 1s",
-			want:    model.Option{},
-			ok:      false,
-		},
-		{
-			name:    "2 opts limit",
-			payload: "1 2",
-			want:    model.Option{},
-			ok:      false,
-		},
-		{
-			name:    "0 opt",
-			payload: "",
-			want:    model.Option{},
-			ok:      false,
-		},
-		{
-			name:    "3 opts",
-			payload: "1s 12 profile",
-			want:    model.Option{},
-			ok:      false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, ok := parseTopAndBottomPayload(tt.payload)
-			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("parseTopAndBottomPayload() got = %v, want %v", got, tt.want)
-			}
-			if ok != tt.ok {
-				t.Errorf("parseTopAndBottomPayload() ok = %v, want %v", ok, tt.ok)
-			}
-		})
-	}
-}
-
-func Test_parseProfilePayload(t *testing.T) {
-	tests := []struct {
-		name    string
-		payload string
-		want    model.Option
+		want    Option
 		ok      bool
 	}{
 		{
 			name:    "1 profile opt",
 			payload: "profile",
-			want:    model.Option{Profile: "profile"},
+			want:    Option{Profile: "profile"},
 			ok:      true,
 		},
 		{
 			name:    "1 duration opt",
 			payload: "1s",
-			want:    model.Option{Date: time.Now().Truncate(time.Hour * 24)},
+			want:    Option{Date: time.Now().Truncate(time.Hour * 24)},
 			ok:      true,
 		},
 		{
 			name:    "2 opts profile 1st",
 			payload: "profile 1s",
-			want:    model.Option{Date: time.Now().Truncate(time.Hour * 24), Profile: "profile"},
+			want:    Option{Date: time.Now().Truncate(time.Hour * 24), Profile: "profile"},
 			ok:      true,
 		},
 		{
 			name:    "2 opts profile 2nd",
 			payload: "1s profile",
-			want:    model.Option{Date: time.Now().Truncate(time.Hour * 24), Profile: "profile"},
+			want:    Option{Date: time.Now().Truncate(time.Hour * 24), Profile: "profile"},
 			ok:      true,
 		},
 		{
 			name:    "2 opts duration incorrect",
 			payload: "profile 1ss",
-			want:    model.Option{},
+			want:    Option{},
 			ok:      false,
 		},
 		{
 			name:    "2 opts profile incorrect",
 			payload: "1s 1s",
-			want:    model.Option{},
+			want:    Option{},
 			ok:      false,
 		},
 		{
 			name:    "2 opts profile incorrect",
 			payload: "1s 10",
-			want:    model.Option{},
+			want:    Option{},
 			ok:      false,
 		},
 		{
 			name:    "0 opt",
 			payload: "",
-			want:    model.Option{},
+			want:    Option{},
 			ok:      false,
 		},
 		{
 			name:    "3 opts",
 			payload: "1s profile 12",
-			want:    model.Option{},
+			want:    Option{},
 			ok:      false,
 		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			got, ok := parseProfilePayload(tt.payload)
+			got, ok := NewProfileOption(tt.payload)
 			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("parseProfilePayload() got = %v, want %v", got, tt.want)
+				t.Errorf("getProfilePayload() got = %v, want %v", got, tt.want)
 			}
 			if ok != tt.ok {
-				t.Errorf("parseProfilePayload() ok = %v, want %v", ok, tt.ok)
+				t.Errorf("getProfilePayload() ok = %v, want %v", ok, tt.ok)
+			}
+		})
+	}
+}
+
+func TestGetRatingOption(t *testing.T) {
+	tests := []struct {
+		name    string
+		payload string
+		want    Option
+		ok      bool
+	}{
+		{
+			name:    "1 duration opt",
+			payload: "1s",
+			want:    Option{Date: time.Now().Truncate(time.Hour * 24)},
+			ok:      true,
+		},
+		{
+			name:    "1 limit opt",
+			payload: "3",
+			want:    Option{Limit: 3},
+			ok:      true,
+		},
+		{
+			name:    "2 opts",
+			payload: "1s 12",
+			want:    Option{Date: time.Now().Truncate(time.Hour * 24), Limit: 12},
+			ok:      true,
+		},
+		{
+			name:    "2 opts duration",
+			payload: "1s 1s",
+			want:    Option{},
+			ok:      false,
+		},
+		{
+			name:    "2 opts limit",
+			payload: "1 2",
+			want:    Option{},
+			ok:      false,
+		},
+		{
+			name:    "0 opt",
+			payload: "",
+			want:    Option{},
+			ok:      false,
+		},
+		{
+			name:    "3 opts",
+			payload: "1s 12 profile",
+			want:    Option{},
+			ok:      false,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got, ok := NewRatingOption(tt.payload)
+			if !reflect.DeepEqual(got, tt.want) {
+				t.Errorf("parseTopAndBottomPayload() got = %v, want %v", got, tt.want)
+			}
+			if ok != tt.ok {
+				t.Errorf("parseTopAndBottomPayload() ok = %v, want %v", ok, tt.ok)
 			}
 		})
 	}
