@@ -2,23 +2,34 @@ package postgres
 
 import (
 	"context"
+	"github.com/alexboor/lbx-telebot/internal/model"
 	"time"
 )
 
 // Count increment messages counter in storage by given value for user in the chat
-func (s *Storage) Count(ctx context.Context, uid, cid int64, dt time.Time, val int) error {
+func (s *Storage) Count(ctx context.Context, uid, cid int64, dt time.Time, count model.Count) error {
 	query := `
-insert into word_count (user_id, chat_id, date, val)
-values ($1, $2, $3, $4)
+insert into counting (user_id, chat_id, date, word, reply, forward, media, sticker, message)
+values ($1, $2, $3, $4, $5, $6, $7, $8, $9)
 on conflict (user_id, chat_id, date)
-    do update set val = word_count.val + excluded.val`
+    do update set word    = counting.word + excluded.word,
+                  reply   = counting.reply + excluded.reply,
+                  forward = counting.forward + excluded.forward,
+                  media   = counting.media + excluded.media,
+                  sticker = counting.sticker + excluded.sticker,
+                  message = counting.message + excluded.message`
 
 	_, err := s.Pool.Exec(
 		ctx, query,
-		uid, // $1
-		cid, // $2
-		dt,  // $3
-		val, // $4
+		uid,           // $1
+		cid,           // $2
+		dt,            // $3
+		count.Word,    // $4
+		count.Reply,   // $5
+		count.Forward, // $6
+		count.Media,   // $7
+		count.Sticker, // $8
+		count.Message, // $9
 	)
 	return err
 }
