@@ -1,7 +1,7 @@
 package migration
 
 var Migrations = []string{
-	wordCount,
 	profile,
 	event, eventParticipant,
+	count, countMessages,
 }
